project completed
This project contains a Java application that combines the nearest neighbor algorithm and the genetic algorithm for the Traveling Salesman Problem (TSP). TSP addresses the challenge of finding the shortest route between specified cities for a salesman. This application reads city coordinates of different sizes, finds the most suitable route based on this data, and displays the result along with the optimal cost value. It was developed using the Java programming language in the Eclipse IDE.